move getImage to own utility
