+++ packages/client/hmi-client/src/assets/utils.ts
@@ -1,0 +1,16 @@
+function getImage(fileName: string) {
+	try {
+		const modules = import.meta.glob('@/assets/images/**/*.{png,svg}', {
+			eager: true,
+			import: 'default'
+		});
+		const moduleKeys = Object.keys(modules);
+		const fileSrc = moduleKeys.find((key) => key.endsWith(fileName));
+		return fileSrc ? (modules[fileSrc] as string) : '';
+	} catch (err) {
+		console.debug('getImage', err);
+		return null;
+	}
+}
+
+export default getImage;
