+++ packages/client/hmi-client/src/page/project/components/tera-project-configuration-modal.vue
@@ -85,10 +85,7 @@
 							:value="`0${i}`"
 						/>
 						<label :for="`thumbnail-0${i}`">
-							<img
-								:src="`src/assets/images/project-thumbnails/0${i}.png`"
-								:alt="`Thumbnail image 0{$i}`"
-							/>
+							<img :src="getImage(`project-thumbnails/0${i}.png`)" :alt="`Thumbnail image 0{$i}`" />
 						</label>
 					</li>
 				</ul>
@@ -124,6 +121,7 @@ import { cloneDeep, isEmpty } from 'lodash';
 import { useRouter } from 'vue-router';
 import { RouteName } from '@/router/routes';
 import { Project } from '@/types/Types';
+import getImage from '@/assets/utils';
 
 const props = defineProps<{
 	confirmText: string;
